src/requirements.txt
auto_mix_prep==0.2.0
beautifulsoup4==4.12.3
langchain==0.3.15
langchain_chroma==0.2.0
langchain_community==0.3.15
langchain_core==0.3.31
langchain_google_genai==2.0.9
langchain_groq==0.2.3
langchain_ollama==0.2.2
langchain_openai==0.3.2
python-dotenv==1.0.1
python_magic==0.4.27
Requests==2.32.3
tqdm==4.67.1
